Move tests to test_connections.py
